is-lnu-service/src/test/java/org/lnu/is/extractor/marriedtype/MarriedTypeParametersExtractorTest.java
package org.lnu.is.extractor.marriedtype;

import static org.junit.Assert.assertEquals;

import java.util.HashMap;
import java.util.Map;

import org.junit.Test;
import org.lnu.is.domain.marriedtype.MarriedType;

public class MarriedTypeParametersExtractorTest {

	private MarriedTypeParametersExtractor unit = new MarriedTypeParametersExtractor();
	
	@Test
	public void testGetParameters() throws Exception {
		// Given
		String name = "Extract Me All";
		String abbrName = "EML";

		MarriedType entity = new MarriedType();
		entity.setName(name);
		entity.setAbbrName(abbrName);

		Map<String, Object> expected = new HashMap<String, Object>();
		expected.put("name", name);
		expected.put("abbrName", abbrName);

		// When
		Map<String, Object> actual = unit.getParameters(entity);

		// Then
		assertEquals(expected, actual);
	}

}
